♻️ refactor: useGuideFaqExpand 훅 추출 및 두 화면 공유 (expand·answer 캐시·detail fetch)
